requirements.prod.txt
gunicorn==23.0.0
whitenoise==6.7.0
django-minify-html==1.9.0
